adding support for "ci skip" indicator in commits.
The one assumption here is that the commit is on "master" (which certainly isn't super flexible...maybe more to come soon!). If a commit message has the defined "skip" command, then the hook will be skipped. (If this hook is being used as pre-push, this will work for all commits in the push: if all commits are have the "skip" message, then the hook will be skipped; if at least one of the messages have the "skip" message, then the hook will run).